requirements/local.txt
# Local development dependencies go here
-r _base.txt
coverage
Flask-Testing
nose
sniffer
nosegrowl2
sure
